changes for switch to vmware for wayland testing
switched to vmware from virt-manager/qemu for wayland testing and changes for the same
to backup and virt-* scripts

also changed borg repository password to be different from its remote login password

+++ .local/bin/virt-stop.sh
@@ -2,12 +2,19 @@
 
 # cgroup-lite qemu-system-x86
 
-for service in sshd.service ssh.service nmbd.service smbd.service nmb.service smb.service nfs-server.service nfs-kernel-server.service rpcbind.service virtlogd-admin.socket virtlockd-admin.socket libvirtd-admin.socket libvirtd.socket virtlogd.service virtlogd.socket virtlockd.socket libvirtd-ro.socket libvirt-guests.service libvirtd.service; do
-  if sudo systemctl list-unit-files $service; then
-    sudo systemctl stop $service 2>/dev/null
-    sudo systemctl disable $service 2>/dev/null
-    sudo killall `echo $service | sed 's/\..*//'` 2>/dev/null
-  fi
-done
+if [ -x /etc/init.d/vmware ]; then
+  sudo /etc/init.d/vmware-USBArbitrator stop
+  sudo /etc/init.d/vmware stop
+  sudo rmmod vmmon vmnet 2>/dev/null
+  sudo rmmod vmw_vsock_vmci_transport vmw_vmci vsock
+else
+  for service in sshd.service ssh.service nmbd.service smbd.service nmb.service smb.service nfs-server.service nfs-kernel-server.service rpcbind.service virtlogd-admin.socket virtlockd-admin.socket libvirtd-admin.socket libvirtd.socket virtlogd.service virtlogd.socket virtlockd.socket libvirtd-ro.socket libvirt-guests.service libvirtd.service; do
+    if sudo systemctl list-unit-files $service; then
+      sudo systemctl stop $service 2>/dev/null
+      sudo systemctl disable $service 2>/dev/null
+      sudo killall `echo $service | sed 's/\..*//'` 2>/dev/null
+    fi
+  done
 
-#sudo rmmod kvm_intel kvm
+  #sudo rmmod kvm_intel kvm
+fi
